+++ CHANGELOG.md
@@ -21,6 +21,14 @@ section here.
 
 ---
 
+## [1.2.0] - Unreleased
+
+### .PAS generated file fix
+
+- The .pas file generator `generate-delphi-compiler-versions-pas` was tweaked to
+produce Delphi 2 code but that introduced a bug when Exit(true); was split across
+two lines. Fixed in #17
+
 ## [1.1.0] - 2026-03-06
 
 Platform support matrix doc + Forward Compatibility in INC + filenames standardized